requirements.txt
allure-pytest==2.8.12
allure-python-commons==2.8.12
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
entrypoints==0.3
deepdiff==4.3.2
jsonschema==3.2.0
Faker==4.0.2
flake8==3.7.9
idna==2.9
importlib-metadata==1.6.0
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyHamcrest==2.0.2
pyparsing==2.4.6
pytest==5.4.1
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
text-unidecode==1.3
urllib3==1.25.8
wcwidth==0.1.9
zipp==3.1.0
